Change fonts, tweak layout, and refactor class
I've been unsatisfied with the text in my CV for a while, as it's too
light.  I once tried to use Equity as the body text, which I used in
my Ph.D thesis, but it felt too narrow.  So now that I have the other
fonts in the MB Type Library I have redone it with Valkyrie for the
body text, Heliotrope for the headings, and Century Supra for my name.

I've also tweaked the layout, increasing heading spacing, changing
font sizes, and moving the section rule to below the heading.  It
currently takes up a little more space than the old CV, but it'll
shrink if need be.

FInally, I've removed the outdated LICENCE file.